Ajustando Editar

+++ agendamento-virtual/src/app/list-agendamento/list-agendamento.component.ts
@@ -11,12 +11,9 @@ import { AgendamentoService } from '../services/agendamento.service';
 export class ListAgendamentoComponent implements OnInit {
 
   agendamento: Agendamento[] = [];
-  agendamento1: Agendamento;
-
 
   constructor(private agendamentoService: AgendamentoService,
     public datePipe: DatePipe) {
-    this.agendamento1 = {} as Agendamento;
   }
 
   ngOnInit(): void {
@@ -29,15 +26,8 @@ export class ListAgendamentoComponent implements OnInit {
     });
   }
 
-  getAgendamento(agendamento: Agendamento) {
-    this.agendamentoService.getAgendamentoId(agendamento).subscribe(x => {
-      this.agendamento1 = x;
-      console.log(x)
-    })
-  }
-
   delete(agendamento: Agendamento) {
-    if (agendamento.id != null) {
+    if (agendamento.id != null && agendamento.id != undefined) {
       if (confirm(`Deseja remover o agendamento do dia ${this.datePipe.transform(agendamento.dia, 'dd/MM/yyyy')}?`)) {
         this.agendamentoService.deleteAgendamento(agendamento).subscribe(() => {
           return this.listar();
